Renames basic_fields to base_fields.
