+++ imagine/__init__.py
@@ -5,7 +5,7 @@ from .likelihoods.simple_likelihood import SimpleLikelihood
 from .fields.field_factory import GeneralFieldFactory
 from .fields.field import GeneralField
 from .fields.grid import *
-from .fields.basic_fields import *
+from .fields.base_fields import *
 
 from .observables.observable_dict import Measurements, Simulations, Covariances, Masks
 from .simulators.simulator import Simulator
@@ -28,11 +28,3 @@ from .observables import dataset
 
 # customized modules
 from .simulators.hammurabi import Hammurabi
-#from .fields.breg_lsa.hamx_field import BregLSA
-#from .fields.breg_lsa.hamx_factory import BregLSAFactory
-#from .fields.brnd_es.hamx_field import BrndES
-#from .fields.brnd_es.hamx_factory import BrndESFactory
-#from .fields.cre_analytic.hamx_field import CREAna
-#from .fields.cre_analytic.hamx_factory import CREAnaFactory
-#from .fields.tereg_ymw16.hamx_field import TEregYMW16
-#from .fields.tereg_ymw16.hamx_factory import TEregYMW16Factory
