+++ imagine/fields/__init__.py
@@ -1,4 +1,4 @@
 from .field import GeneralField
 from .field_factory import GeneralFieldFactory
 from .grid import UniformGrid
-from .basic_fields import *
+from .base_fields import *
